add annotations and refactor routes
