+++ packages/send/backend/src/routes/users.ts
@@ -32,6 +32,27 @@ import { requireJWT } from '../middleware';
 
 const router: Router = Router();
 
+/**
+ * @swagger
+ * /users/me:
+ *   get:
+ *     summary: Get current user information
+ *     description: Retrieves the logged-in user from the current session
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: User information retrieved successfully
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 user:
+ *                   type: object
+ *       404:
+ *         description: User not found
+ */
 router.get(
   '/me',
   requireJWT,
@@ -53,6 +74,25 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/publickey/{id}:
+ *   get:
+ *     summary: Get user's public key
+ *     description: Retrieves the public key for a specific user
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Public key retrieved successfully
+ *       404:
+ *         description: Public key not found
+ */
 router.get(
   '/publickey/:id',
   addErrorHandling(USER_ERRORS.PUBLIC_KEY_NOT_FOUND),
@@ -66,6 +106,29 @@ router.get(
 // Update user's public key
 // TODO: decide whether this should just be a "profile update" function
 // that can take any of the following: email, publicKey, avatar...
+/**
+ * @swagger
+ * /users/publickey:
+ *   post:
+ *     summary: Update user's public key
+ *     description: Updates the public key for the authenticated user
+ *     security:
+ *       - bearerAuth: []
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               publicKey:
+ *                 type: string
+ *     responses:
+ *       200:
+ *         description: Public key updated successfully
+ *       400:
+ *         description: Profile update failed
+ */
 router.post(
   '/publickey',
   requireJWT,
@@ -89,6 +152,20 @@ router.post(
   })
 );
 
+/**
+ * @swagger
+ * /users/folders:
+ *   get:
+ *     summary: Get user's folders
+ *     description: Retrieves all folders for the authenticated user
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: Folders retrieved successfully
+ *       404:
+ *         description: Folders not found
+ */
 router.get(
   '/folders',
   requireJWT,
@@ -104,6 +181,27 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/lookup/{email}:
+ *   get:
+ *     summary: Lookup user by email
+ *     description: Retrieves user information by email address
+ *     security:
+ *       - bearerAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: email
+ *         required: true
+ *         schema:
+ *           type: string
+ *         description: User's email address
+ *     responses:
+ *       200:
+ *         description: User found
+ *       404:
+ *         description: User not found
+ */
 router.get(
   '/lookup/:email',
   requireJWT,
@@ -119,6 +217,20 @@ router.get(
 
 // TODO: shift userId to session and out of req.params
 
+/**
+ * @swagger
+ * /users/login:
+ *   post:
+ *     summary: User login
+ *     description: Authenticates and logs in a user
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: Login successful
+ *       404:
+ *         description: Login failed
+ */
 router.post(
   '/login',
   requireJWT,
@@ -135,6 +247,32 @@ router.post(
   })
 );
 
+/**
+ * @swagger
+ * /users:
+ *   post:
+ *     summary: Create new user
+ *     description: Creates a new user with the provided information
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               publicKey:
+ *                 type: string
+ *               email:
+ *                 type: string
+ *               tier:
+ *                 type: string
+ *                 enum: [FREE, PRO, ENTERPRISE]
+ *     responses:
+ *       201:
+ *         description: User created successfully
+ *       400:
+ *         description: User creation failed
+ */
 router.post(
   '/',
   addErrorHandling(USER_ERRORS.USER_NOT_CREATED),
@@ -164,7 +302,25 @@ router.post(
   })
 );
 
-// All containers, regardless of type
+/**
+ * @swagger
+ * /users/{userId}/containers:
+ *   get:
+ *     summary: Get all user containers
+ *     description: Retrieves all containers for a specific user
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Containers retrieved successfully
+ *       404:
+ *         description: Containers not found
+ */
 router.get(
   '/:userId/containers',
   addErrorHandling(USER_ERRORS.FOLDERS_NOT_FOUND),
@@ -175,6 +331,25 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/{userId}/conversations:
+ *   get:
+ *     summary: Get user conversations
+ *     description: Retrieves all conversations for a specific user
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Conversations retrieved successfully
+ *       404:
+ *         description: Conversations not found
+ */
 router.get(
   '/:userId/conversations',
   addErrorHandling(USER_ERRORS.FOLDERS_NOT_FOUND),
@@ -188,6 +363,25 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/{userId}/activity:
+ *   get:
+ *     summary: Get user activity
+ *     description: Retrieves recent activity for a specific user
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Activity retrieved successfully
+ *       404:
+ *         description: Activity not found
+ */
 router.get(
   '/:userId/activity',
   addErrorHandling(USER_ERRORS.HISTORY_NOT_FOUND),
@@ -201,6 +395,25 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/{userId}/folders/sharedByUser:
+ *   get:
+ *     summary: Get folders shared by user
+ *     description: Retrieves all folders that a user has shared with others
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Shared folders retrieved successfully
+ *       404:
+ *         description: Shared folders not found
+ */
 router.get(
   '/:userId/folders/sharedByUser',
   addErrorHandling(USER_ERRORS.SHARED_FOLDERS_NOT_FOUND),
@@ -219,6 +432,25 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/{userId}/folders/sharedWithUser:
+ *   get:
+ *     summary: Get folders shared with user
+ *     description: Retrieves all folders that have been shared with a user
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Received folders retrieved successfully
+ *       404:
+ *         description: Received folders not found
+ */
 router.get(
   '/:userId/folders/sharedWithUser',
   addErrorHandling(USER_ERRORS.RECEIVED_FOLDERS_NOT_FOUND),
@@ -233,7 +465,25 @@ router.get(
   })
 );
 
-// Get invitations for user
+/**
+ * @swagger
+ * /users/{userId}/invitations:
+ *   get:
+ *     summary: Get user invitations
+ *     description: Retrieves all invitations for a specific user
+ *     parameters:
+ *       - in: path
+ *         name: userId
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Invitations retrieved successfully
+ *       404:
+ *         description: Invitations not found
+ */
 router.get(
   '/:userId/invitations',
   addErrorHandling(USER_ERRORS.INVITATIONS_NOT_FOUND),
@@ -244,6 +494,40 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/{id}/backup:
+ *   post:
+ *     summary: Create user backup
+ *     description: Creates a backup for a specific user
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               keys:
+ *                 type: string
+ *               keypair:
+ *                 type: string
+ *               keystring:
+ *                 type: string
+ *               salt:
+ *                 type: string
+ *     responses:
+ *       200:
+ *         description: Backup created successfully
+ *       400:
+ *         description: Backup failed
+ */
 router.post(
   '/:id/backup',
   addErrorHandling(USER_ERRORS.BACKUP_FAILED),
@@ -258,6 +542,35 @@ router.post(
   })
 );
 
+/**
+ * @swagger
+ * /users/backup:
+ *   post:
+ *     summary: Create authenticated user backup
+ *     description: Creates a backup for the authenticated user
+ *     security:
+ *       - bearerAuth: []
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               keys:
+ *                 type: string
+ *               keypair:
+ *                 type: string
+ *               keystring:
+ *                 type: string
+ *               salt:
+ *                 type: string
+ *     responses:
+ *       200:
+ *         description: Backup created successfully
+ *       400:
+ *         description: Backup failed
+ */
 router.post(
   '/backup',
   requireJWT,
@@ -273,6 +586,20 @@ router.post(
   })
 );
 
+/**
+ * @swagger
+ * /users/backup:
+ *   get:
+ *     summary: Get authenticated user backup
+ *     description: Retrieves the backup for the authenticated user
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: Backup retrieved successfully
+ *       404:
+ *         description: Backup not found
+ */
 router.get(
   '/backup',
   requireJWT,
@@ -284,6 +611,27 @@ router.get(
   })
 );
 
+/**
+ * @swagger
+ * /users/{id}/backup:
+ *   get:
+ *     summary: Get user backup
+ *     description: Retrieves the backup for a specific user
+ *     security:
+ *       - bearerAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: User ID
+ *     responses:
+ *       200:
+ *         description: Backup retrieved successfully
+ *       404:
+ *         description: Backup not found
+ */
 router.get(
   '/:id/backup',
   requireJWT,
