+++ packages/send/backend/src/trpc/users.ts
@@ -26,15 +26,98 @@ import { router, publicProcedure as trpc } from '../trpc';
 import { isAuthed, requirePublicLogin, useEnvironment } from './middlewares';
 
 export const usersRouter = router({
+  /**
+   * @openapi
+   * /trpc/getUser:
+   *   get:
+   *     tags:
+   *       - Users
+   *     summary: Get current user ID
+   *     description: Returns the ID of the currently authenticated user
+   *     responses:
+   *       200:
+   *         description: User ID
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 user:
+   *                   type: number
+   *                   description: ID of the current user
+   */
   getUser: trpc.query(({ ctx }) => {
     return { user: Number(ctx.user.id) };
   }),
 
+  /**
+   * @openapi
+   * /trpc/getUserData:
+   *   get:
+   *     tags:
+   *       - Users
+   *     summary: Get current user data
+   *     description: Returns the complete user data for the currently authenticated user
+   *     security:
+   *       - bearerAuth: []
+   *     responses:
+   *       200:
+   *         description: User data
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 userData:
+   *                   type: object
+   *                   description: Complete user data object
+   */
   getUserData: trpc.use(isAuthed).query(async ({ ctx }) => {
     const userData = await getUserById(Number(ctx.user.id));
     return { userData: userData };
   }),
 
+  /**
+   * @openapi
+   * /trpc/settings:
+   *   get:
+   *     tags:
+   *       - Users
+   *     summary: Get API settings and compatibility
+   *     description: Returns API version and compatibility information
+   *     parameters:
+   *       - in: query
+   *         name: input
+   *         schema:
+   *           type: object
+   *           properties:
+   *             version:
+   *               type: string
+   *               description: Client version to check compatibility against
+   *     responses:
+   *       200:
+   *         description: API settings and compatibility information
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 apiVersion:
+   *                   type: string
+   *                   description: Current API version
+   *                 compatibility:
+   *                   type: object
+   *                   properties:
+   *                     resolvedCompatibility:
+   *                       type: boolean
+   *                       description: Whether compatibility has been resolved
+   *                     result:
+   *                       type: string
+   *                       description: Compatibility result
+   *                 clientVersion:
+   *                   type: string
+   *                   description: Client version that was checked
+   */
   settings: trpc
     .input(
       z.object({
@@ -56,6 +139,27 @@ export const usersRouter = router({
       return { apiVersion, compatibility, clientVersion };
     }),
 
+  /**
+   * @openapi
+   * /trpc/onLoginFinished:
+   *   get:
+   *     tags:
+   *       - Users
+   *     summary: Subscribe to login completion events
+   *     description: Returns a stream of login completion events
+   *     parameters:
+   *       - in: query
+   *         name: input
+   *         schema:
+   *           type: object
+   *           properties:
+   *             name:
+   *               type: string
+   *               description: Name of the login event
+   *     responses:
+   *       200:
+   *         description: Stream of login completion events
+   */
   onLoginFinished: trpc
     .input(
       z.object({
@@ -77,11 +181,28 @@ export const usersRouter = router({
       }
     }),
 
-  // ==========================
-  // DO NOT USE IN PRODUCTION
-  // ==========================
-
-  // This mutation allows authed users to reset their passphrase
+  /**
+   * @openapi
+   * /trpc/resetKeys:
+   *   post:
+   *     tags:
+   *       - Users
+   *     summary: Reset user keys (Development Only)
+   *     description: Resets user keys and deletes all containers and uploads (Development Only)
+   *     security:
+   *       - bearerAuth: []
+   *     responses:
+   *       200:
+   *         description: Keys reset successfully
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 success:
+   *                   type: boolean
+   *                   description: Whether the reset was successful
+   */
   resetKeys: trpc
     .use(isAuthed)
     .use((props) => useEnvironment(props, ['stage', 'development']))
@@ -107,7 +228,39 @@ export const usersRouter = router({
       }
     }),
 
-  /* This mutation should not be used in production */
+  /**
+   * @openapi
+   * /trpc/userLogin:
+   *   post:
+   *     tags:
+   *       - Users
+   *     summary: User login (Development Only)
+   *     description: Authenticates a user with email and password (Development Only)
+   *     requestBody:
+   *       required: true
+   *       content:
+   *         application/json:
+   *           schema:
+   *             type: object
+   *             properties:
+   *               email:
+   *                 type: string
+   *                 description: User's email address
+   *               password:
+   *                 type: string
+   *                 description: User's password
+   *     responses:
+   *       200:
+   *         description: Login successful
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 state:
+   *                   type: string
+   *                   description: Login state token
+   */
   userLogin: trpc
     .use(requirePublicLogin)
     .input(
@@ -144,7 +297,42 @@ export const usersRouter = router({
       }
     }),
 
-  /* This mutation should not be used in production */
+  /**
+   * @openapi
+   * /trpc/registerUser:
+   *   post:
+   *     tags:
+   *       - Users
+   *     summary: Register new user (Development Only)
+   *     description: Creates a new user account (Development Only)
+   *     requestBody:
+   *       required: true
+   *       content:
+   *         application/json:
+   *           schema:
+   *             type: object
+   *             properties:
+   *               email:
+   *                 type: string
+   *                 description: User's email address
+   *               password:
+   *                 type: string
+   *                 description: User's password
+   *     responses:
+   *       200:
+   *         description: User registered successfully
+   *         content:
+   *           application/json:
+   *             schema:
+   *               type: object
+   *               properties:
+   *                 id:
+   *                   type: number
+   *                   description: ID of the newly created user
+   *                 state:
+   *                   type: string
+   *                   description: Login state token
+   */
   registerUser: trpc
     .use(requirePublicLogin)
     .input(
